Improved switching between submodes
Created the Submode ui class for the submodes to inherit from.
This allows the switch_to_{submode} commands to all be available
from all submodes.
Also made an improvement to the printing of RuleCalls and gave them id numbers
for quickly referencing which is intended when using the switch command.